refactor: split app.go into domain-specific files
app.go had grown past 5200 lines. Split it by domain following the
existing app_file_transfer.go pattern, moving functions verbatim with
no behavior changes:

- app_terminal.go: session lifecycle, jump-host tunnels, monitor panel,
  session output logging, terminal history, serial/shell helpers
- app_database.go: SQL/Redis/MongoDB/Elasticsearch methods, grouped by
  backend with session helpers
- app_container.go: container and Kubernetes methods
- app_remote_desktop.go: RDP window controls and X11 desktop connect
- app_ai.go: chat completion backends, model management, skills

Merge the standalone app_credentials.go back into app.go and drop four
unused Anthropic SSE type declarations.

Public API unchanged: wails3 generate bindings produces no diff.
